Change to unit64 for bytes count

+++ pcap-convert/app/include/pcapconvert/PcapConvertStats.h
@@ -2,6 +2,7 @@
 
 // Standard includes
 #include <cstddef>
+#include <cstdint>
 
 namespace pcapconvert
 {
@@ -10,9 +11,9 @@ namespace pcapconvert
   {
   private:
     // Data
-    size_t bytesDropped = 0;
-    size_t bytesIn = 0;
-    size_t bytesOut = 0;
+    uint64_t bytesDropped = 0;
+    uint64_t bytesIn = 0;
+    uint64_t bytesOut = 0;
     size_t dnsModifiedPackets = 0;
     size_t droppedPackets = 0;
     size_t writtenPackets = 0;
@@ -23,18 +24,18 @@ namespace pcapconvert
     virtual ~PcapConvertStats();
 
     // Getters & setters
-    size_t getBytesDropped() const { return bytesDropped; }
-    size_t getBytesIn() const { return bytesIn; }
-    size_t getBytesOut() const { return bytesOut; }
+    uint64_t getBytesDropped() const { return bytesDropped; }
+    uint64_t getBytesIn() const { return bytesIn; }
+    uint64_t getBytesOut() const { return bytesOut; }
     size_t getDnsModifiedPackets() const { return dnsModifiedPackets; }
     size_t getDroppedPackets() const { return droppedPackets; }
     size_t getWrittenPackets() const { return writtenPackets; }
 
     // Business logic
-    void incrementBytesIn(size_t value) { bytesIn += value; };
+    void incrementBytesIn(uint64_t value) { bytesIn += value; };
     void incrementDnsModifiedPackets() { ++dnsModifiedPackets; };
-    void recordDroppedPacket(size_t value);
-    void recordWrittenPacket(size_t value);
+    void recordDroppedPacket(uint64_t value);
+    void recordWrittenPacket(uint64_t value);
   };
 
 } // namespace pcapconvert
